add getAllOrderById method

+++ src/main/java/com/example/contoso/dto/response/MostPopularItemResponse.java
@@ -10,6 +10,6 @@ import lombok.Builder;
 @Builder
 public record MostPopularItemResponse (
         String key,
-        Integer value
+        Long value
 ) {
 }
